// EcomWeb/src/api/client/models/Category.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Product } from './Product';
export type Category = {
    isDeleted?: boolean;
    readonly id?: number;
    readonly name?: string | null;
    readonly description?: string | null;
    readonly products?: Array<Product> | null;
};